返回类型修改!

+++ takin-web-data/src/main/java/io/shulie/takin/web/data/dao/perfomanceanaly/PressureMachineDaoImpl.java
@@ -39,7 +39,7 @@ public class PressureMachineDaoImpl implements PressureMachineDao {
     }
 
     @Override
-    public Integer getCountByIp(String ip) {
+    public Long getCountByIp(String ip) {
         LambdaQueryWrapper<PressureMachineEntity> wrapper = new LambdaQueryWrapper<>();
         wrapper.eq(PressureMachineEntity::getIp, ip);
         return pressureMachineMapper.selectCount(wrapper);
